requirements.txt
faiss-cpu
sentence-transformers
numpy
flask